requirements.txt
python-fasthtml
supabase
python-dotenv